Fully configurable tests now possible
The process described by the design flow chart is now implemented from
end to end, though it is still lacking in many places. Most of the
changes were to the new update thread that threadManagers run, where
batches are distributed to clients. I also cleaned up output quite a
bit, which is important for checking the progress of a running program.
I also added a "log n" command to the controller which shows the most
recent n events from the server.