Enhanced thread lifecycle robustness, added start of the settings module
New clients now exchange a handshake packet with the threadManager to
see just how new they are, and if they are determined to be an old
client reconnecting, the connection handle is redirected to its proper
thread. This functionality is far from done and tasks are not restarted
properly, but it is a start. The settings module is in a similar state,
with much of the groundwork done but a lot of polish still to be done.
The rest of the software also needs to start to use it for it to be
useful. Both of these new features are entirely untested and likely
don't work at all.