Result batching system
Results are now sent to the server in smaller batches than the task
batches to prevent connection problems from interrupting operation.
Client errors are logged to their own file now. Connection threads are
now passed the batch number along with the batch so they will be able to
inform the server about the specific tasks that need to be reassigned if
disconnection occurs.