Started client performance analysis
Added the basis for a function that calculates the portion of work done
by each client, for use in post-run reports or for status updates during
execution. It currently adds up the number of tasks that each client has
done by looking at the batches that were marked as being completed by
that client.