Server displays results received rather than batches
Instead of displaying the number of results in the batches received, the
server now shows how many results it holds plus the number of results
that the connectionThreads haven't yet presented. This quantity is
managed by threadManager and is not integral to anything but the output
on the server, so it is possible that it doesn't get reset properly at
some point.