Added argument checking to controller
It now verifies that the user has entered the proper number and types of
arguments for many commands through use of the new "checkFormat"
function. There are some commands that still need to be done, but this
is a big improvement over it erroring out if the user enters one
incorrect command.